Add version module
